refactor : 계정 생성용 빌더 추가 / 일부 setter 추가
